cleanup lpandarb bot (#1356)
I have been looking over this file closely to spec out the
`targeted_long` and `target_rate_deadzone` logic. This is an interim PR
since I ended up making some superficial changes:
- reorganized use of temporary PoolState objects to reduce the number of
slow `deepcopy` calls
- reorganized arguments & fn placement to be consistent
- extract most of the `action` fn code into new fns
- added docs and type hints
- cleaned up inefficiencies with repeated RPCs
- unified logging format & deleted a time-to-completion log that
required a lot of overhead (I would recommend that we further remove a
bulk of these logs, and move others from `info` to `debug`, but I didn't
want to step on anyone's use case without prior discussion)
- moved some logic to the base class
- added explicit use of the current pool state to avoid accidental race
conditions
- switches use of stdlib `min` and `max` to the `fixedpointmath`
version; there was a type error before that could have resulted in a
loss of precision.
- added a test & fix for a bug where the agent could spend it's money
adding liquidity and then try to spend it again arbing

The existing tests were not changed, and all still pass. None of the
underlying logic should have changed, aside from the bug fix and
avoiding race conditions. However I made changes to when temp variables
were used, so I recommend a close look to catch anything the tests may
have missed.

On my machine, before this PR the `lpandarb_test.py` file took 25.45
seconds to complete, now it takes 24.99 seconds (excluding the two new
tests).

+++ src/agent0/core/base/config/budget.py
@@ -39,9 +39,11 @@ class Budget:
             A sample from a clipped random normal distribution according to the parameters defined at construction
         """
         return FixedPoint(
-            scaled_value=clip(
-                int(np.round(rng.normal(loc=self.mean_wei, scale=self.std_wei))),
-                self.min_wei,
-                self.max_wei,
+            scaled_value=int(
+                clip(
+                    int(np.round(rng.normal(loc=self.mean_wei, scale=self.std_wei))),
+                    self.min_wei,
+                    self.max_wei,
+                )
             )
         )
